Added "manage drivers" feature for Administrator users
--- Resume ---
Added feature that allows Administrator users to manage drivers from a
the manageDrivers page view. The Administrator users are able to view,
add, edit and remove drivers from the drivers table on the database.

--- Details ---
- Added a drivers model that manages CRUD operations on the drivers
table from the database.

- Added manageDrivers page view.

- Added the following functions to Padrera controller:
    · manageDrivers(): initializes the manageDrivers page view.
    · addDriver(): invokes the insertDriver() function from the drivers
        model.
    · editDriver(): invokes the updateDriver() function from the
        drivers model.
    · removeDriver(): invokes the removeDriver() function from the
        drivers model.

- Modified getMenuCamiones() function from utils_helper.php to show
the options "Gestionar Conductores" and "Gestionar Camiones". Only
visible to Administrator users.