Added "block drivers" feature
--- Resume ---
Added a feature that allows Administrator users to block drivers from
the manageDrivers page view. The Administrator users are able to block
drivers until a specified date and write the reason for why the driver
was blocked.

--- Details ---
- Added a "blocked drivers log" model that obtains and inserts "block
logs".

- Added the blockDriver() function to the Pedrera controller.

- Modified manageDrivers page view to include a "status" (estado)
column in the drivers table that tells whether a driver is "active"
(activo) or "blocked" (bloqueado). If the driver is blocked, a button
will appear that allows Administrator users to see the reason for the
block. A block button has also been added to the actions column for
every row in the drivers table.