Added "manage trucks" feature for Administrator users
--- Resume ---
Added feature that allows Administrator users to manage trucks from the
manageTrucks page view. The Administrator users are able to view, add,
edit, remove trucks from the trucks table on the database. Administrator
users are also able to block trucks until a specified date and write
the reason for why the truck was blocked.

--- Details ---
- Modified the trucks model to manage CRUD operations on the trucks
table from the database.

- Added manageTrucks page view.

- Added a "blocked trucks log" model that obtains and inserts "block
logs".

- Added the following functions to Padrera controller:
    · manageTrucks(): initializes the manageTrucks page view.
    · addTruckAdmin(): invokes the insertTruck() function from the
        trucks model.
    · editTruck(): invokes the updateTruck() function from the trucks
        model.
    · removeTruck(): invokes the removeTruck() function from the trucks
        model.
    · blockTruck(): invokes the addLog() function from the "blocked
        trucks log" model.